+++ src/dailyforecastbox.h
@@ -31,7 +31,6 @@ public:
     this->row_height = this->bounding_box.h / DailyForecastBox::row_count;
   }
 
-protected:
   void do_paint() override {
     this->draw_values();
     this->draw_frame();